fix(smrt-svelte): move types to utils.ts for proper TypeScript resolution
Move TimeEntry and ApprovalStatus types from Svelte files to utils.ts
to fix TypeScript module resolution issues in CI environment.
