+++ packages/smrt-svelte/src/components/time/index.ts
@@ -2,25 +2,24 @@
  * Time tracking components - Time entries, approvals, and timesheets
  */
 
-// Types
-export type { ApprovalStatus } from './ApprovalActions.svelte';
 // Components
 export { default as ApprovalActions } from './ApprovalActions.svelte';
 export { default as BulkActions } from './BulkActions.svelte';
 export { default as DurationDisplay } from './DurationDisplay.svelte';
 export { default as RejectDialog } from './RejectDialog.svelte';
-export type { TimeEntry } from './TimeEntryCard.svelte';
 export { default as TimeEntryCard } from './TimeEntryCard.svelte';
 export { default as TimeEntryList } from './TimeEntryList.svelte';
 export { default as TimeSummary } from './TimeSummary.svelte';
 
-// Utilities
+// Types and utilities
 export {
+  type ApprovalStatus,
   type Currency,
   formatCurrency,
   formatDate,
   formatHours,
   formatHoursHHMM,
   statusColors,
+  type TimeEntry,
   type TimeEntryStatus,
 } from './utils.js';
